Renaming ai_sdk.js to generate.js
