+++ sdk/llm/src/generate.js
