+++ sdk/llm/src/index.js
@@ -1,4 +1,4 @@
-export { generateText, generateTextWithStreaming, streamText, generateImage } from './ai_sdk.js';
+export { generateText, generateTextWithStreaming, streamText, generateImage } from './generate.js';
 export { Agent, createMemoryConversationStore } from './agent.js';
 export { loadPrompt } from './prompt/loader.js';
 export { registerProvider, getProviderNames } from './ai_provider.js';
